+++ src/app/dashboard/services/widgets/ASTUE/astue-efficiency.service.ts
@@ -30,6 +30,8 @@ export class AstueEfficiencyService {
 
     public data: BehaviorSubject<IAsEfProduct[]> = new BehaviorSubject<IAsEfProduct[]>(null);
 
+    public productId: string;
+
     private readonly restUrl: string;
 
     get currentUnit(): IAsEfUnitNew {
@@ -53,6 +55,7 @@ export class AstueEfficiencyService {
 
     constructor(private snackbar: SnackBarService, public http: HttpClient, configService: AppConfigService) {
         this.restUrl = configService.restUrl;
+        this.selectionUnit$.subscribe(value => console.log(value));
     }
 
     public isCardOpen(unitName: string): boolean {
@@ -124,6 +127,7 @@ export class AstueEfficiencyService {
                 this.unitsFlowsMap[unit.name] = [];
             }
         });
+        this.selectionUnit$.next([...units]);
         this.selection$.next();
     }