+++ src/app/widgets/CMID/cmid-dynamics-of-metrics-change/models/cmid-dynamics-of-metrics-change.model.ts
@@ -1,0 +1,4 @@
+export interface ICmidDynamicsOfMetricsChangeModel {
+    value: number;
+    date: string;
+}