+++ src/app/widgets/EC/ec-widget-factor-analysis/components/ec-widget-factor-analisys-components/ec-widget-factor-analysis-bar-multi-level/directives/factory-analysis-tooltip.directive.ts
@@ -17,6 +17,13 @@ const POSITIONS_MAP = {
         overlayY: 'center',
         offsetY: 40,
     },
+    top: {
+        originX: 'center',
+        originY: 'center',
+        overlayX: 'center',
+        overlayY: 'center',
+        offsetY: -35,
+    },
 };
 @Directive({
     selector: '[evjFactoryAnalysisTooltip]',
@@ -66,7 +73,8 @@ export class FactoryAnalysisTooltipDirective implements OnInit, OnDestroy {
         tooltipRef.instance.unit = this.unit;
     }
 
-    @HostListener('mouseout')
+    @HostListener('wheel')
+    @HostListener('mouseleave')
     public hide(): void {
         this.overlayRef.detach();
     }