+++ src/app/widgets/EVJ/evj-events-workspace/evj-tasks-event-no-checkbox/evj-tasks-event-no-checkbox.component.ts
@@ -5,8 +5,7 @@ import { IChatMessageWithAttachments } from '../components/evj-chat/evj-chat.com
 @Component({
     selector: 'evj-evj-tasks-event-no-checkbox',
     templateUrl: './evj-tasks-event-no-checkbox.component.html',
-    styleUrls: ['./evj-tasks-event-no-checkbox.component.scss'],
-    changeDetection: ChangeDetectionStrategy.OnPush,
+    styleUrls: ['./evj-tasks-event-no-checkbox.component.scss']
 })
 export class EvjTasksEventNoCheckboxComponent implements OnInit {
     @Input()
@@ -39,11 +38,15 @@ export class EvjTasksEventNoCheckboxComponent implements OnInit {
         return obj ? obj[infoType] : undefined;
     }
 
-    public dateTimePicker(date: Date): void {
-        this.ewService.setDeadlineToEvent(date);
-    }
-
     public onChangeEventDescription(description: string): void {
         this.ewService.event.description = description;
+    }
+
+    public setStartToEvent(value: Date): void {
+        this.ewService.setStartToEvent(value);
+    }
+
+    public setDeadlineToEvent(value: Date): void {
+        this.ewService.setDeadlineToEvent(value);
     }
 }