+++ src/app/widgets/SOU/sou-streams/sou-streams.component.ts
@@ -8,24 +8,52 @@ import {
 } from '@angular/core';
 import { WidgetService } from '@dashboard/services/widget.service';
 import { WidgetPlatform } from '@dashboard/models/@PLATFORM/widget-platform';
+    AfterViewChecked, ChangeDetectorRef
+} from '@angular/core';
+import { WidgetService } from '@dashboard/services/widget.service';
+import { WidgetPlatform } from '@dashboard/models/@PLATFORM/widget-platform';
 
 import { TITLES_OF_TABLE } from '@widgets/SOU/sou-streams/config';
 import { TABLE_CELLS } from '@widgets/SOU/sou-streams/mock';
 import { Overlay, OverlayRef } from '@angular/cdk/overlay';
 import { TemplatePortal } from '@angular/cdk/portal';
+import { TITLES_OF_TABLE } from '@widgets/SOU/sou-streams/config';
+import { TABLE_CELLS } from '@widgets/SOU/sou-streams/mock';
+import { animate, state, style, transition, trigger } from '@angular/animations';
 
 
 @Component({
     selector: 'evj-sou-streams',
     templateUrl: './sou-streams.component.html',
-    styleUrls: ['./sou-streams.component.scss']
+    styleUrls: ['./sou-streams.component.scss'],
+    animations: [
+        trigger(
+            'inOutAnimation',
+            [
+                transition(
+                    ':enter',
+                    [
+                        style({ height: 0, opacity: 0, transform: 'translateY(-593px)' }),
+                        animate('1s ease-out',
+                            style({ height: 593, opacity: 1, transform: 'translateY(0px)' }))
+                    ]
+                ),
+                transition(
+                    ':leave',
+                    [
+                        style({ height: 300, opacity: 1 }),
+                        animate('1s ease-in',
+                            style({ height: 0, opacity: 0 }))
+                    ]
+                )
+            ]
+        )
+    ]
 })
 export class SouStreamsComponent extends WidgetPlatform implements OnInit, AfterViewChecked, AfterViewInit, OnDestroy {
 
     public titlesOfTable: { name: string, bigBlock?: boolean }[] = TITLES_OF_TABLE;
     public tableRows: {} = TABLE_CELLS;
-
-    public isReservoirTrendOpen: boolean = false;
 
     public heightOfTable: string = '400px';
     public heightOfViewPort: string = '335px';
@@ -35,6 +63,9 @@ export class SouStreamsComponent extends WidgetPlatform implements OnInit, After
     @ViewChild('overlayCustom') dialogTemplate: TemplateRef<any>;
     private overlayRef: OverlayRef;
     private portal: TemplatePortal;
+
+    public showWorkspace: boolean = false;
+    public show: boolean = true;
 
     constructor(
         private cdr: ChangeDetectorRef,
@@ -70,7 +101,6 @@ export class SouStreamsComponent extends WidgetPlatform implements OnInit, After
     }
 
     @ViewChild('widget') child: ElementRef;
-
     ngAfterViewChecked(): void {
         this.heightOfTable = this.child.nativeElement.clientHeight - 40 + 'px';
         this.heightOfViewPort = this.child.nativeElement.clientHeight - 115 + 'px';
@@ -85,6 +115,25 @@ export class SouStreamsComponent extends WidgetPlatform implements OnInit, After
         this.cdr.detectChanges();
     }
 
+    get stateName(): string {
+        return this.show ? 'show' : 'hide';
+    }
+
+    protected dataHandler(ref: unknown): void {
+    }
+
+    toggle(): void {
+        this.show = !this.show;
+    }
+
+    public addNewOperation(): void {
+        this.showWorkspace = true;
+        this.toggle();
+        const height: number = this.showWorkspace ? 593 : 0;
+        this.heightOfTable = this.child.nativeElement.clientHeight - 40 - height + 'px';
+        this.heightOfViewPort = this.child.nativeElement.clientHeight - 115 - height + 'px';
+        this.cdr.detectChanges();
+    }
     public openReservoirTrend(): void {
         this.isReservoirTrendOpen = true;
         console.log(window);