+++ src/app/widgets/SOU/sou-streams/sou-streams.module.ts
@@ -11,6 +11,13 @@ import { TrendsComponent } from './smart-trend/components/trends/trends.componen
 import { TrendTableComponent } from './smart-trend/components/trend-table/trend-table.component';
 import { DragDropModule } from '@angular/cdk/drag-drop';
 
+import { SouStreamsComponent } from '@widgets/SOU/sou-streams/sou-streams.component';
+import { SharedModule } from '@shared/shared.module';
+import { MatSelectModule } from '@angular/material/select';
+import { AngularSvgIconModule } from 'angular-svg-icon';
+import { ScrollingModule } from '@angular/cdk/scrolling';
+import { MatRadioModule } from '@angular/material/radio';
+import { SouWorkspaceModule } from '@widgets/SOU/sou-workspace/sou-workspace.module';
 
 
 @NgModule({